Adding the ability to run the server file from anywhere. Moving CheckboxCollection and WidgetLogger components to a components directory.
