+++ TwitchVotingServer/ServerGUI/components/WidgetLogger.py
