+++ TwitchVotingServer/config/game_effects/DSR.ini
@@ -1,17 +1,17 @@
 [STAT_EFFECTS]
-one_hp = False
+one_hp = True
 
 [SETTINGS_EFFECTS]
-lower_draw_distance = False
+lower_draw_distance = True
 blind = True
 
 [GUI_EFFECTS]
-disable_hud = False
+disable_hud = True
 
 [HEAD_EFFECTS]
-big_head = False
-tiny_head = False
+big_head = True
+tiny_head = True
 
 [ACTION_EFFECTS]
-sliding = False
+sliding = True
 warp_to_bonfire = True
